Time: 0 ms (100.00%) | Memory: 8.6 MB (20.70%) - LeetSync

+++ 7-reverse-integer/reverse-integer.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int reverse(int x) {
-        long rev=0;
+        long long rev=0;
         int i=0;
         while(x!=0){
             int last=x%10;
